Rename ParticipantEntry namespace dir
https://github.com/eSolutions-EMS/endurance-management-system/issues/158
