+++ src/Witness/EMS.Witness/Pages/Witness/Components/SendFormComponent.razor
@@ -1,7 +1,7 @@
 ﻿@inherits StatefulComponent
 @using Core.Application.Services;
 @using Core.Domain.AggregateRoots.Manager
-@using Core.Domain.AggregateRoots.Manager.Aggregates.ParticipantEntries;
+@using Core.Domain.AggregateRoots.Manager.Aggregates.Participants;
 @using Core.Models;
 @using EMS.Witness.Common;
 @using EMS.Witness.Services;
